+++ packages/custom-upload/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @naive-potato-ui/custom-upload
 
+## 1.0.14
+
+### Patch Changes
+
+- change upload parse
+
 ## 1.0.13
 
 ### Patch Changes